+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - `RP_CLIENT_TYPE` configuration variable, by @HardNorth
 - `RP_CONNECT_TIMEOUT` and `RP_READ_TIMEOUT` configuration variables, by @HardNorth
 ### Changed
-- Client version updated on [5.5.1](https://github.com/reportportal/client-Python/releases/tag/5.5.1), by @HardNorth
+- Client version updated on [5.5.2](https://github.com/reportportal/client-Python/releases/tag/5.5.2), by @HardNorth
 
 ## [5.2.2]
 ### Changed